server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.0
pydantic-settings==2.6.1
httpx==0.25.1
python-multipart==0.0.6
PyJWT==2.8.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.3
boto3==1.34.0
openai==1.3.7
python-dotenv==1.0.0
pytest==8.3.5
